tools/pyproject.toml
[project]
name = "archive-tools"
version = "0.1.0"
description = "CLI tools for korikov.cc archive site"
requires-python = ">=3.14"
dependencies = [
    "pyzotero",
    "pydantic>=2.0",
    "click",
    "pyyaml",
    "transliterate>=1.10.2",
    "hugo>=0.153.0",
]

[project.scripts]
fetch = "fetch:main"
generate = "generate:main"
validate = "validate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.djlint]
profile = "hugo"
# H006: img width/height (Hugo-rendered images)
# H016/H030/H031: title/meta in partials (not full pages)
# H021: inline styles (Yandex Metrica)
# H025: orphan </a> inside Go template printf strings
# D018/J018: Django/Jinja URL patterns (not Hugo)
ignore = "D018,H006,H016,H021,H025,H030,H031,J018"

[tool.pyright]
pythonVersion = "3.14"
typeCheckingMode = "strict"

[dependency-groups]
dev = [
    "djlint>=1.36.4",
    "playwright>=1.58.0",
    "pytest-playwright>=0.7.2",
    "ruff>=0.14.14",
    "syrupy>=5.1.0",
    "yamllint>=1.38.0",
]
